chore(); replace authentication hooks with user store
