+++ ui/apps/akasha/src/components/app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { I18nextProvider } from 'react-i18next';
-import { useGetLogin, useRootComponentProps } from '@akashaorg/ui-awf-hooks';
+import { useAkashaStore, useRootComponentProps } from '@akashaorg/ui-awf-hooks';
 import { router } from './app-routes';
 import { RouterProvider } from '@tanstack/react-router';
 import { useApolloClient } from '@apollo/client';
@@ -14,7 +14,9 @@ declare module '@tanstack/react-router' {
 const SocialApp: React.FC<unknown> = () => {
   const { getTranslationPlugin, baseRouteName } = useRootComponentProps();
   const apolloClient = useApolloClient();
-  const login = useGetLogin();
+  const {
+    data: { authenticatedDID },
+  } = useAkashaStore();
 
   return (
     <I18nextProvider i18n={getTranslationPlugin().i18n}>
@@ -22,7 +24,7 @@ const SocialApp: React.FC<unknown> = () => {
         router={router({
           baseRouteName,
           apolloClient,
-          authenticatedDID: login.data?.id ?? '',
+          authenticatedDID,
         })}
       />
     </I18nextProvider>
