+++ ui/apps/akasha/src/components/pages/entry-page/beam-page.tsx
@@ -9,8 +9,8 @@ import {
   hasOwn,
   mapBeamEntryData,
   mapReflectEntryData,
+  useAkashaStore,
   useAnalytics,
-  useGetLogin,
   useNsfwToggling,
   useRootComponentProps,
 } from '@akashaorg/ui-awf-hooks';
@@ -34,12 +34,14 @@ const BeamPage: React.FC<BeamPageProps> = props => {
   const { beamId, beamStream, beam } = props;
   const { t } = useTranslation('app-akasha-integration');
   const { navigateToModal, getTranslationPlugin } = useRootComponentProps();
-  const { data, loading: authenticating } = useGetLogin();
+  const {
+    data: { authenticatedDID, isAuthenticating: authenticating },
+  } = useAkashaStore();
   const { showNsfw } = useNsfwToggling();
   const [analyticsActions] = useAnalytics();
   const wrapperRef = useRef(null);
   const navigate = useNavigate();
-  const isLoggedIn = !!data?.id;
+  const isLoggedIn = !!authenticatedDID;
 
   /**
    * Check the current moderation status of the beam
