+++ ui/apps/extensions/src/components/pages/info-page.tsx
@@ -7,7 +7,7 @@ import {
 import Stack from '@akashaorg/design-system-core/lib/components/Stack';
 import AppInfo from '@akashaorg/design-system-components/lib/components/AppInfo';
 import { useTranslation } from 'react-i18next';
-import { hasOwn, transformSource, useGetLogin } from '@akashaorg/ui-awf-hooks';
+import { hasOwn, transformSource, useAkashaStore } from '@akashaorg/ui-awf-hooks';
 import { useGetAppReleaseByIdQuery } from '@akashaorg/ui-awf-hooks/lib/generated/apollo';
 
 type InfoPageProps = {
@@ -15,17 +15,11 @@ type InfoPageProps = {
 };
 
 const InfoPage: React.FC<InfoPageProps> = ({ appId }) => {
-  // const { getRoutingPlugin } = useRootComponentProps();
-
-  // const navigateTo = getRoutingPlugin().navigateTo;
-
   const { t } = useTranslation('app-extensions');
-
-  const { data } = useGetLogin();
-  const isLoggedIn = !!data?.id;
-
-  // const network = useCurrentNetwork(isLoggedIn).data;
-
+  const {
+    data: { authenticatedDID },
+  } = useAkashaStore();
+  const isLoggedIn = !!authenticatedDID;
   const { data: appReleaseInfoReq, error } = useGetAppReleaseByIdQuery({
     variables: { id: appId },
     skip: !isLoggedIn,
@@ -35,46 +29,6 @@ const InfoPage: React.FC<InfoPageProps> = ({ appId }) => {
     appReleaseInfoReq.node && hasOwn(appReleaseInfoReq?.node, 'applicationID')
       ? appReleaseInfoReq.node
       : null;
-
-  // const author = appReleaseInfo.application.author.akashaProfile;
-
-  // const handleAuthorEthAddressClick = (ethAddress: string) => {
-  //   if (network) {
-  //     window.open(
-  //       `https://${network}.etherscan.io/address/${ethAddress}`,
-  //       '_blank',
-  //       'noreferrer noopener',
-  //     );
-  //   } else {
-  //     window.open(`https://etherscan.io/address/${ethAddress}`, '_blank', 'noreferrer noopener');
-  //   }
-  // };
-
-  // const handleAuthorClick = (author: { id: string }) => {
-  //   navigateTo?.({
-  //     appName: '@akashaorg/app-profile',
-  //     getNavigationUrl: routes => `${routes.rootRoute}/${author.id}`,
-  //   });
-  // };
-
-  // @TODO update with new hooks when available
-  // const installedAppsReq = useGetAllInstalledApps(isLoggedIn);
-
-  // const isInstalled = React.useMemo(() => {
-  //   if (installedAppsReq.data) {
-  //     const installedAppsIds = installedAppsReq.data.map(app => app.id);
-  //     return installedAppsIds.includes(appId);
-  //   }
-  // }, [installedAppsReq.data, appId]);
-
-  // const releasesInfoReq = useGetAppsReleasesQuery(
-  //   { last: 10 },
-  //   {
-  //     select: resp => resp.akashaAppReleaseIndex.edges,
-  //   },
-  // );
-
-  // const releases = releasesInfoReq.data.map(release => release.node);
 
   const developers = appReleaseInfo?.application?.contributors?.map(contributor => {
     const avatarImg = contributor.akashaProfile?.avatar;
@@ -92,7 +46,6 @@ const InfoPage: React.FC<InfoPageProps> = ({ appId }) => {
           type="script-error"
           title={t('There was an error loading the app info')}
           details={t('We cannot show this app right now')}
-          // devDetails={appReleaseInfoReq.error.message}
         />
       )}
       {appReleaseInfo && (
