+++ ui/apps/search/src/components/pages/search-page.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   type ITag,
@@ -6,19 +6,15 @@ import {
   AnalyticsCategories,
   type Profile,
 } from '@akashaorg/typings/lib/ui';
-
 import routes, { SETTINGS } from '../../routes';
-
 import {
   useEntryNavigation,
   useAnalytics,
   useRootComponentProps,
   transformSource,
-  useGetLogin,
+  useAkashaStore,
 } from '@akashaorg/ui-awf-hooks';
-
 import { SearchTagsResult } from '@akashaorg/typings/lib/sdk/graphql-types';
-
 import EntryCardRenderer from './entry-renderer';
 import Card from '@akashaorg/design-system-core/lib/components/Card';
 import Stack from '@akashaorg/design-system-core/lib/components/Stack';
@@ -63,8 +59,10 @@ const SearchPage: React.FC<SearchPageProps> = props => {
   const { t } = useTranslation('app-search');
   const { getRoutingPlugin, navigateToModal } = useRootComponentProps();
 
-  const { data } = useGetLogin();
-  const isLoggedIn = !!data?.id;
+  const {
+    data: { authenticatedDID },
+  } = useAkashaStore();
+  const isLoggedIn = !!authenticatedDID;
 
   const showLoginModal = () => {
     navigateToModal({ name: 'login' });
