+++ ui/apps/settings-app/src/components/pages/nsfw-option.tsx
@@ -5,13 +5,15 @@ import Stack from '@akashaorg/design-system-core/lib/components/Stack';
 import Text from '@akashaorg/design-system-core/lib/components/Text';
 import Toggle from '@akashaorg/design-system-core/lib/components/Toggle';
 import PageLayout from './base-layout';
-import { useGetLogin, useNsfwToggling, useRootComponentProps } from '@akashaorg/ui-awf-hooks';
+import { useAkashaStore, useNsfwToggling, useRootComponentProps } from '@akashaorg/ui-awf-hooks';
 
 const NsfwOption: React.FC = () => {
   const { t } = useTranslation('app-settings-ewa');
 
-  const { data, loading } = useGetLogin();
-  const isLoggedIn = !!data?.id;
+  const {
+    data: { authenticatedDID, isAuthenticating },
+  } = useAkashaStore();
+  const isLoggedIn = !!authenticatedDID;
 
   const { getRoutingPlugin, uiEvents } = useRootComponentProps();
   const routingPlugin = useRef(getRoutingPlugin());
@@ -19,7 +21,7 @@ const NsfwOption: React.FC = () => {
 
   const { showNsfw, toggleShowNsfw } = useNsfwToggling();
 
-  if (!isLoggedIn && !loading) {
+  if (!isLoggedIn && !isAuthenticating) {
     // if not logged in, redirect to homepage
     routingPlugin.current?.navigateTo?.({
       appName: '@akashaorg/app-akasha-integration',
