+++ ui/widgets/trending/src/components/trending-widget-component.tsx
@@ -6,7 +6,12 @@ import TrendingWidgetLoadingCard from '@akashaorg/design-system-components/lib/c
 import Text from '@akashaorg/design-system-core/lib/components/Text';
 import Card from '@akashaorg/design-system-core/lib/components/Card';
 import { useTranslation } from 'react-i18next';
-import { useRootComponentProps, getFollowList, hasOwn, useGetLogin } from '@akashaorg/ui-awf-hooks';
+import {
+  useRootComponentProps,
+  getFollowList,
+  hasOwn,
+  useAkashaStore,
+} from '@akashaorg/ui-awf-hooks';
 import {
   useGetInterestsStreamQuery,
   useGetInterestsByDidQuery,
@@ -19,9 +24,10 @@ import { useGetIndexingDID } from '@akashaorg/ui-awf-hooks/lib/use-settings';
 
 const TrendingWidgetComponent: React.FC<unknown> = () => {
   const { t } = useTranslation('ui-widget-trending');
-  const { data } = useGetLogin();
-  const isLoggedIn = !!data?.id;
-  const authenticatedDID = data?.id;
+  const {
+    data: { authenticatedDID },
+  } = useAkashaStore();
+  const isLoggedIn = !!authenticatedDID;
   const { plugins, uiEvents, logger, navigateToModal } = useRootComponentProps();
   const navigateTo = plugins['@akashaorg/app-routing']?.routing?.navigateTo;
